requirements.txt
Flask==3.1.1
Werkzeug==3.1.3
python-docx==0.8.11
